AWSM - Framework - Fix paths to previous storm_days
Fix logic that sets the paths to the previous storm_day.nc.
There was a path error which caused the new file not be found and the
initial file (day one) to be re-used throughout a configured time span.
This fixes the path and also resets the filepath once the file can not
be found.