CLI - Fix `run_awsm_daily` date access
Parsed values after the config gets parsed need to access the `.cfg`
property instead of the `raw_cfg` since that is only a list of raw
values. The parsed values are also already `datetime.datetime` objects
and no need to cast as string to parse again as pd Timestamp objects.