AWSM CLI - Add ability to run full water year at once
Expand the CLI options to allow configuring a full water year in the .ini
file and run without having to call the first day separately.
When giving the combination of an .ini file and the `no_previous` flag
it is assumed that the first day in the .ini file has not previous output
data.